separate files for training aquire and handling
